Connected to the API
Recieving data works now. had to fiddle around with SSL certificates, but it works now. Auth token is not visible within auth.php because that is sensitive information. this means the user needs to provide it themself. (If this were to go on a hosting platform it would be available there, but since this is a github repository i'd rather not expose it.)